+++ src/context/ConfigContext.tsx
@@ -4,6 +4,7 @@ import { FillError } from "@dialob/fill-api";
 export interface ConfigContextType {
   errors: (items: FillError[]) => React.ReactElement;
   description: (text: string) => React.ReactElement;
+  breadCrumbs: (items: string[], canNavigate: boolean, activeItem?: string) => React.ReactElement;
 }
 
 const context = React.createContext<ConfigContextType>({ 
@@ -13,6 +14,10 @@ const context = React.createContext<ConfigContextType>({
   },
   description: (text: string) => {
     console.log("description has no impl.", text);
+    return (<></>);
+  },
+  breadCrumbs: (items: string[]) => {
+    console.log("breadCrumbs has no impl.", items);
     return (<></>);
   }
 });