+++ src/index.ts
@@ -17,6 +17,7 @@ export {TextBox, TextBoxProps} from './components/TextBox';
 export {TimeField, TimeFieldProps} from './components/TimeField';
 export {Survey, SurveyProps} from './components/Survey';
 export {SurveyGroup, SurveyGroupProps} from './components/SurveyGroup';
+export {BreadCrumbs, BreadCrumbsProps} from './components/BreadCrumbs';
 
 export {MaterialDialob, MaterialDialobProps} from './MaterialDialob';
 export {DefaultView, DefaultViewProps} from './DefaultView';